+++ test/_env/srv/handlers/main.js
@@ -109,6 +109,34 @@ module.exports = (srv) => {
     return text + "-" + req.context.user.id;
   });
 
+  srv.on("triggerCustomRoleEvent", async (req) => {
+    const ID = req.data.ID;
+    const text = req.data.text + req.data.num;
+    await srv.emit("customRoleIncludeEvent", { ID, text, user: req.context.user.id, role: "admin" });
+    await srv.emit("customRoleExcludeEvent", { ID, text, user: req.context.user.id, role: "abc" });
+    return text + "-" + req.context.user.id;
+  });
+
+  srv.on("triggerCustomRoleDynamicEvent", async (req) => {
+    const ID = req.data.ID;
+    const text = req.data.text + req.data.num;
+    await srv.emit("customRoleIncludeDynamicEvent", {
+      ID,
+      text,
+      user: req.context.user.id,
+      role: "admin",
+      flag: ["admin", req.context.user.id],
+    });
+    await srv.emit("customRoleExcludeDynamicEvent", {
+      ID,
+      text,
+      user: req.context.user.id,
+      role: "abc",
+      flag: [req.context.user.id],
+    });
+    return text + "-" + req.context.user.id;
+  });
+
   srv.on("triggerCustomContextHeaderEvent", async (req) => {
     const ID = req.data.ID;
     const text = req.data.text + req.data.num;
@@ -137,6 +165,18 @@ module.exports = (srv) => {
         wsUserInclude: [],
         userExclude: [],
         wsUserExclude: [],
+        role: {
+          include: [],
+          exclude: [],
+        },
+        wsRole: {
+          include: [],
+          exclude: [],
+        },
+        roleInclude: [],
+        wsRoleInclude: [],
+        roleExclude: [],
+        wsRoleExclude: [],
         currentUser: {
           include: req.data.num === 0,
           exclude: req.data.num === 1,